Changed FileContentHandler to avoid deleting the temporary file a second time. The FileInputStream.finalize method calls close, which previously logged a duplicate message about deleting the temporary file.

+++ projects/otex-core/source/java/com/google/enterprise/connector/otex/FileContentHandler.java
@@ -80,9 +80,16 @@ class FileContentHandler implements ContentHandler {
                         try {
                             super.close();
                         } finally {
-                            if (LOGGER.isLoggable(Level.FINER))
-                                LOGGER.finer("DELETE: " + temporaryFile);
-                            temporaryFile.delete();
+                            // close is called again from the finalizer,
+                            // so check whether the file still exists.
+                            // No synchronization is needed, because the
+                            // finalizer can't be running while another
+                            // thread is using this object.
+                            if (temporaryFile.exists()) {
+                                if (LOGGER.isLoggable(Level.FINER))
+                                    LOGGER.finer("DELETE: " + temporaryFile);
+                                temporaryFile.delete();
+                            }
                         }
                     }
                 };
